refactor(#318): split run.ts monolith into focused modules
Extract 3,865-line run.ts into 4 focused modules under src/lib/workflow/:

- worktree-manager.ts (1,428 lines): Worktree lifecycle, freshness checks,
  PR creation, rebase operations
- phase-executor.ts (493 lines): Phase execution with retry, QA verdict
  parsing, cold-start handling
- phase-mapper.ts (209 lines): Label-to-phase detection, workflow parsing,
  phase determination
- batch-executor.ts (888 lines): Batch execution, dependency sorting,
  issue logging, RunOptions

run.ts reduced to 960 lines (orchestration + re-exports).

All 1,511 tests pass. No behavior changes.

Fixes #318
